+++ test/test-osvvm.tcl
@@ -4,7 +4,7 @@
 # Note: tcllib is required for OSVVM
 
 # known good snapshot
-set OsvvmLibraries_tag "2023.04"
+set OsvvmLibraries_tag "2023.05"
 set OsvvmDir [file join $::env(HOME) .cache nvc "OsvvmLibraries-${OsvvmLibraries_tag}"]
 
 source $OsvvmDir/Scripts/StartNVC.tcl